Follow guidelines in hooks, remove redundant credits.
