+++ mass_mailing_unique/README.rst
@@ -46,11 +46,6 @@ mass_mailing_unique%0Aversion:%20
 Credits
 =======
 
-Images
-------
-
-* Odoo Community Association: `Icon <https://github.com/OCA/maintainer-tools/blob/master/template/module/static/description/icon.svg>`_.
-
 Contributors
 ------------
